requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.36
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic
pydantic-settings
python-dotenv==1.0.0
pillow
aiofiles==23.2.1
google-generativeai==0.3.2
httpx==0.26.0
alembic==1.13.1
email-validator==2.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
websockets==12.0
alembic
